requirements.txt
flask>=3.1.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0